Refactor numeric option coercion
